+++ CHANGELOG.md
@@ -29,7 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   been removed.
 - **Breaking**: The cdt queue types, `DT_QUEUE`, `Dtqueue`, `_Dtqueue` have been
   removed.
-- **Breaking**: The cdt alias `struct _dthold_s` has been removed.
+- **Breaking**: The cdt aliases `struct _dthold_s`, `struct _dtdata_s` have been
+  removed.
 
 ### Fixed